refactor(desktop): extract agent lifecycle actions into a hook
UserProfilePanel.tsx crossed the 1025-line file-size guard after the
Restart action landed. Move the start/stop/restart handlers into
useAgentLifecycleActions so the panel shrinks below the limit.
